crud movie ok

+++ src/main/java/com/filmFlix/project_filmFlix/projections/MovieProjection.java
@@ -3,6 +3,6 @@ package com.filmFlix.project_filmFlix.projections;
 public interface MovieProjection {
     String getTitle();
     String getSub_title();
-    Integer getMovie_year();
+    Long getMovie_year();
     String getImg_url();
 }
